Value changed triggered only on control change
Currently this control behaves differently to default iOS controls such as UISlider.  When AXRatingView value is set from code it will currently trigger UIControlEventValueChanged; this is not how UISlider works.

UISlider will only trigger value changed event when user interacts with the control to change its value.

UIControl class reference documentation states the following.
"UIControlEventValueChanged: A touch dragging or otherwise manipulating a control, causing it to emit a series of different values."

This commit ensures that UIControlEventValueChanged is only triggered when the value is changed through user interaction.

Also given the potential to break backwards compatibility for some users, if you choose to accept this change it should probably be part of a minor release and not a dot release.